# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Healthcare Appointment System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientBody"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientProfile"
  /patients/{id}/dashboard:
    get:
      operationId: getPatientDashboard
      tags: [patients]
      summary: Patient dashboard data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDashboard"
...
  /reports:
    get:
      operationId: getReports
      tags: [reports]
      summary: List medical reports
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: reportType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListResponse"

    post:
      operationId: createReport
      tags: [reports]
      summary: Upload report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /prescriptions:
    get:
      operationId: getPrescriptions
      tags: [prescriptions]
      summary: List prescriptions
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: doctorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Prescriptions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescriptionListResponse"

    post:
      operationId: createPrescription
      tags: [prescriptions]
      summary: Create prescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrescriptionBody"
      responses:
        "201":
          description: Prescription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
...
  /leave-blocks:
    get:
      operationId: getLeaveBlocks
      tags: [doctors]
      summary: Get doctor leave blocks
      parameters:
        - name: doctorId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leave blocks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveBlockListResponse"

    post:
      operationId: createLeaveBlock
      tags: [doctors]
      summary: Create leave block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaveBlockBody"
      responses:
        "201":
          description: Leave block created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveBlock"
  /leave-blocks/{id}:
    delete:
      operationId: deleteLeaveBlock
      tags: [doctors]
      summary: Delete leave block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /emergency:
    get:
      operationId: getEmergencyCases
      tags: [emergency]
      summary: List emergency cases
      responses:
        "200":
          description: Emergency cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyCaseListResponse"

    post:
      operationId: createEmergencyCase
      tags: [emergency]
      summary: Create emergency case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmergencyCaseBody"
      responses:
        "201":
          description: Emergency case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyCase"
...
  /emergency/ambulance:
    post:
      operationId: requestAmbulance
      tags: [emergency]
      summary: Request ambulance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmbulanceRequestBody"
      responses:
        "200":
          description: Ambulance dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AmbulanceResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        errors:
          type: array
          items:
            type: string
      required:
        - success
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Patient:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        dob:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        bloodGroup:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        chronicConditions:
          type: ["string", "null"]
      required:
        - id
        - userId
        - name
        - email
    PatientProfile:
      allOf:
        - $ref: "#/components/schemas/Patient"
    UpdatePatientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        dob:
          type: string
        gender:
          type: string
        bloodGroup:
          type: string
        address:
          type: string
        emergencyContact:
          type: string
        allergies:
          type: string
        chronicConditions:
          type: string
        avatarUrl:
          type: string
    PatientDashboard:
      type: object
      properties:
        upcomingAppointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        recentReports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        stats:
          $ref: "#/components/schemas/PatientStats"
        healthSummary:
          $ref: "#/components/schemas/HealthSummary"
        has24HourAlert:
          type: boolean
      required:
        - upcomingAppointments
        - recentReports
        - stats
        - healthSummary
        - has24HourAlert
    PatientStats:
      type: object
      properties:
        total:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer
        upcoming:
          type: integer
      required:
        - total
        - completed
        - cancelled
        - upcoming
    HealthSummary:
      type: object
      properties:
        bloodGroup:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        chronicConditions:
          type: ["string", "null"]
      required: []
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        slotId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        tokenNumber:
          type: integer
        queuePosition:
          type: integer
        estimatedWaitMins:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, no-show]
        reason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        paymentAmount:
          type: number
        createdAt:
          type: string
        patientName:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        doctorSpecialization:
          type: ["string", "null"]
      required:
        - id
        - patientId
        - doctorId
        - slotId
        - date
        - startTime
        - endTime
        - tokenNumber
        - queuePosition
        - estimatedWaitMins
        - status
        - paymentStatus
        - paymentAmount
        - createdAt
...
    Report:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: ["integer", "null"]
        appointmentId:
          type: ["integer", "null"]
        filePath:
          type: ["string", "null"]
        fileType:
          type: ["string", "null"]
        reportType:
          type: string
          enum: [Blood Test, X-Ray, MRI, CT Scan, Ultrasound, Prescription, Other]
        title:
          type: string
        notes:
          type: ["string", "null"]
        uploadedAt:
          type: string
        doctorName:
          type: ["string", "null"]
      required:
        - id
        - patientId
        - reportType
        - title
        - uploadedAt
    ReportListResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
      required:
        - reports
    CreateReportBody:
      type: object
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        appointmentId:
          type: integer
        reportType:
          type: string
        title:
          type: string
        notes:
          type: string
      required:
        - patientId
        - reportType
        - title
    Medicine:
      type: object
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        durationDays:
          type: integer
      required:
        - name
        - dosage
        - frequency
        - durationDays
    Prescription:
      type: object
      properties:
        id:
          type: integer
        appointmentId:
          type: integer
        doctorId:
          type: integer
        patientId:
          type: integer
        medicines:
          type: array
          items:
            $ref: "#/components/schemas/Medicine"
        instructions:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        createdAt:
          type: string
        doctorName:
          type: ["string", "null"]
        patientName:
          type: ["string", "null"]
      required:
        - id
        - appointmentId
        - doctorId
        - patientId
        - medicines
        - createdAt
    PrescriptionListResponse:
      type: object
      properties:
        prescriptions:
          type: array
          items:
            $ref: "#/components/schemas/Prescription"
      required:
        - prescriptions
    CreatePrescriptionBody:
      type: object
      properties:
        appointmentId:
          type: integer
        doctorId:
          type: integer
        patientId:
          type: integer
        medicines:
          type: array
          items:
            $ref: "#/components/schemas/Medicine"
        instructions:
          type: string
        validUntil:
          type: string
      required:
        - appointmentId
        - doctorId
        - patientId
        - medicines
...
    LeaveBlock:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        date:
          type: string
        isFullDay:
          type: boolean
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
      required:
        - id
        - doctorId
        - date
        - isFullDay
    LeaveBlockListResponse:
      type: object
      properties:
        leaveBlocks:
          type: array
          items:
            $ref: "#/components/schemas/LeaveBlock"
      required:
        - leaveBlocks
    CreateLeaveBlockBody:
      type: object
      properties:
        doctorId:
          type: integer
        date:
          type: string
        isFullDay:
          type: boolean
        startTime:
          type: string
        endTime:
          type: string
        reason:
          type: string
      required:
        - doctorId
        - date
        - isFullDay
    EmergencyCase:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: string
        patientName:
          type: string
        age:
          type: integer
        issueType:
          type: string
        severity:
          type: string
          enum: [red, orange, green]
        conscious:
          type: string
          enum: [yes, no]
        location:
          type: string
        status:
          type: string
          enum: [open, assigned, resolved]
        assignedDoctorId:
          type: ["integer", "null"]
        createdAt:
          type: string
        assignedDoctorName:
          type: ["string", "null"]
        autoBookedAppointmentId:
          type: ["integer", "null"]
      required:
        - id
        - caseId
        - patientName
        - age
        - issueType
        - severity
        - conscious
        - location
        - status
        - createdAt
    EmergencyCaseListResponse:
      type: object
      properties:
        cases:
          type: array
          items:
            $ref: "#/components/schemas/EmergencyCase"
      required:
        - cases
    CreateEmergencyCaseBody:
      type: object
      properties:
        patientName:
          type: string
        age:
          type: integer
        issueType:
          type: string
        severity:
          type: string
          enum: [red, orange, green]
        conscious:
          type: string
          enum: [yes, no]
        location:
          type: string
      required:
        - patientName
        - age
        - issueType
        - severity
        - conscious
        - location
...
    AmbulanceRequestBody:
      type: object
      properties:
        location:
          type: string
        caseId:
          type: string
      required:
        - location
    AmbulanceResponse:
      type: object
      properties:
        driverName:
          type: string
        vehicleNumber:
          type: string
        etaMinutes:
          type: integer
        phone:
          type: string
      required:
        - driverName
        - vehicleNumber
        - etaMinutes
        - phone
...