# lib/api-zod/src/generated/types/emergencyCase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Healthcare Appointment System API
 * OpenAPI spec version: 0.1.0
 */
import type { EmergencyCaseConscious } from "./emergencyCaseConscious";
import type { EmergencyCaseSeverity } from "./emergencyCaseSeverity";
import type { EmergencyCaseStatus } from "./emergencyCaseStatus";

export interface EmergencyCase {
  id: number;
  caseId: string;
  patientName: string;
  age: number;
  issueType: string;
  severity: EmergencyCaseSeverity;
  conscious: EmergencyCaseConscious;
  location: string;
  status: EmergencyCaseStatus;
  /** @nullable */
  assignedDoctorId?: number | null;
  createdAt: string;
  /** @nullable */
  assignedDoctorName?: string | null;
  /** @nullable */
  autoBookedAppointmentId?: number | null;
}