# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Healthcare Appointment System API
 * OpenAPI spec version: 0.1.0
 */
import type { Medicine } from "./medicine";

export interface Prescription {
  id: number;
  appointmentId: number;
  doctorId: number;
  patientId: number;
  medicines: Medicine[];
  /** @nullable */
  instructions?: string | null;
  /** @nullable */
  validUntil?: string | null;
  createdAt: string;
  /** @nullable */
  doctorName?: string | null;
  /** @nullable */
  patientName?: string | null;
}